Reorganize routes into directories
