+++ endpoints/instantiations/metadata_/route.py
